pyproject.toml
[project]
name = "jagirsathi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apify-client>=2.4.0",
    "langchain-groq>=1.1.1",
    "mcp[cli]>=1.26.0",
    "openai>=2.16.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.53.1",
]
